+++ src/main/scala/org/izolotov/crawler/parser/PlanetaSportParser.scala
@@ -17,7 +17,7 @@ object PlanetaSportParser extends Parser[Product]{
       val title = doc.select("h1.productHeader").first.ownText
       val brand = doc.select("h1.productHeader b").text
       val category = doc.select("li.breadcrumbs__item a").asScala.drop(1).dropRight(1).map(e => e.text())//.drop(2).map(e => e.text())
-      val oldPrice: Option[Int] = doc.select("span.oldPrise").text() match {
+      val oldPrice: Option[Float] = doc.select("span.oldPrise").text() match {
         case "" => None
         case somePrice => Some(Util.parsePrice(somePrice))
       }
